+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/Schema/AccessorSparseIndices.cs
@@ -18,8 +18,15 @@ namespace GLTFast.Schema
         /// The index of the bufferView with sparse indices.
         /// Referenced bufferView can't have ARRAY_BUFFER or ELEMENT_ARRAY_BUFFER target.
         /// </summary>
-        [JsonPropertyName("bufferView")]
-        public uint BufferView { get; set; }
+        [JsonIgnore]
+        public int BufferView { get; set; } = Constants.UnsetIndex;
+
+        [JsonPropertyName("bufferView"), JsonInclude]
+        internal int? BufferViewSerialized
+        {
+            get => BufferView < 0 ? null : BufferView;
+            set => BufferView = value ?? Constants.UnsetIndex;
+        }
 
         /// <summary>
         /// The offset relative to the start of the bufferView in bytes. Must be aligned.
@@ -56,7 +63,10 @@ namespace GLTFast.Schema
         internal void GltfSerialize(JsonWriter writer)
         {
             writer.AddObject();
-            writer.AddProperty("bufferView", BufferView);
+            if (BufferView >= 0)
+            {
+                writer.AddProperty("bufferView", BufferView);
+            }
             writer.AddProperty("componentType", ComponentType);
             if (ByteOffset >= 0)
             {