+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/Schema/AccessorSparseValues.cs
@@ -18,8 +18,15 @@ namespace GLTFast.Schema
         /// The index of the bufferView with sparse values.
         /// Referenced bufferView can't have ARRAY_BUFFER or ELEMENT_ARRAY_BUFFER target.
         /// </summary>
-        [JsonPropertyName("bufferView")]
-        public uint BufferView { get; set; }
+        [JsonIgnore]
+        public int BufferView { get; set; } = Constants.UnsetIndex;
+
+        [JsonPropertyName("bufferView"), JsonInclude]
+        internal int? BufferViewSerialized
+        {
+            get => BufferView < 0 ? null : BufferView;
+            set => BufferView = value ?? Constants.UnsetIndex;
+        }
 
         /// <summary>
         /// The offset relative to the start of the bufferView in bytes. Must be aligned.
@@ -47,7 +54,10 @@ namespace GLTFast.Schema
         internal void GltfSerialize(JsonWriter writer)
         {
             writer.AddObject();
-            writer.AddProperty("bufferView", BufferView);
+            if (BufferView >= 0)
+            {
+                writer.AddProperty("bufferView", BufferView);
+            }
             if (ByteOffset >= 0)
             {
                 writer.AddProperty("byteOffset", ByteOffset);